Restructure language server code for testability
